refacot displayTags provider
